move hunyuan stuff to image_text
