+++ src/diffusers/models/embeddings/__init__.py
@@ -1,13 +1,13 @@
 from .combined import (
     CombinedTimestepLabelEmbeddings,
     CombinedTimestepTextProjEmbeddings,
-    HunyuanCombinedTimestepTextSizeStyleEmbedding,
-    HunyuanDiTAttentionPool,
     LabelEmbedding,
     PixArtAlphaCombinedTimestepSizeEmbeddings,
 )
 from .image_text import (
     AttentionPooling,
+    HunyuanCombinedTimestepTextSizeStyleEmbedding,
+    HunyuanDiTAttentionPool,
     ImageHintTimeEmbedding,
     ImagePositionalEmbeddings,
     ImageProjection,
