+++ src/diffusers/models/embeddings/combined.py
@@ -1,6 +1,5 @@
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 
 
 class LabelEmbedding(nn.Module):
@@ -114,108 +113,5 @@ class PixArtAlphaCombinedTimestepSizeEmbeddings(nn.Module):
             conditioning = timesteps_emb + torch.cat([resolution_emb, aspect_ratio_emb], dim=1)
         else:
             conditioning = timesteps_emb
-
-        return conditioning
-
-
-class HunyuanDiTAttentionPool(nn.Module):
-    # Copied from https://github.com/Tencent/HunyuanDiT/blob/cb709308d92e6c7e8d59d0dff41b74d35088db6a/hydit/modules/poolers.py#L6
-
-    def __init__(self, spacial_dim: int, embed_dim: int, num_heads: int, output_dim: int = None):
-        super().__init__()
-        self.positional_embedding = nn.Parameter(torch.randn(spacial_dim + 1, embed_dim) / embed_dim**0.5)
-        self.k_proj = nn.Linear(embed_dim, embed_dim)
-        self.q_proj = nn.Linear(embed_dim, embed_dim)
-        self.v_proj = nn.Linear(embed_dim, embed_dim)
-        self.c_proj = nn.Linear(embed_dim, output_dim or embed_dim)
-        self.num_heads = num_heads
-
-    def forward(self, x):
-        x = x.permute(1, 0, 2)  # NLC -> LNC
-        x = torch.cat([x.mean(dim=0, keepdim=True), x], dim=0)  # (L+1)NC
-        x = x + self.positional_embedding[:, None, :].to(x.dtype)  # (L+1)NC
-        x, _ = F.multi_head_attention_forward(
-            query=x[:1],
-            key=x,
-            value=x,
-            embed_dim_to_check=x.shape[-1],
-            num_heads=self.num_heads,
-            q_proj_weight=self.q_proj.weight,
-            k_proj_weight=self.k_proj.weight,
-            v_proj_weight=self.v_proj.weight,
-            in_proj_weight=None,
-            in_proj_bias=torch.cat([self.q_proj.bias, self.k_proj.bias, self.v_proj.bias]),
-            bias_k=None,
-            bias_v=None,
-            add_zero_attn=False,
-            dropout_p=0,
-            out_proj_weight=self.c_proj.weight,
-            out_proj_bias=self.c_proj.bias,
-            use_separate_proj_weight=True,
-            training=self.training,
-            need_weights=False,
-        )
-        return x.squeeze(0)
-
-
-class HunyuanCombinedTimestepTextSizeStyleEmbedding(nn.Module):
-    def __init__(
-        self,
-        embedding_dim,
-        pooled_projection_dim=1024,
-        seq_len=256,
-        cross_attention_dim=2048,
-        use_style_cond_and_image_meta_size=True,
-    ):
-        super().__init__()
-        from .image_text import PixArtAlphaTextProjection
-        from .timestep import TimestepEmbedding, Timesteps
-
-        self.time_proj = Timesteps(num_channels=256, flip_sin_to_cos=True, downscale_freq_shift=0)
-        self.timestep_embedder = TimestepEmbedding(in_channels=256, time_embed_dim=embedding_dim)
-
-        self.size_proj = Timesteps(num_channels=256, flip_sin_to_cos=True, downscale_freq_shift=0)
-
-        self.pooler = HunyuanDiTAttentionPool(
-            seq_len, cross_attention_dim, num_heads=8, output_dim=pooled_projection_dim
-        )
-
-        # Here we use a default learned embedder layer for future extension.
-        self.use_style_cond_and_image_meta_size = use_style_cond_and_image_meta_size
-        if use_style_cond_and_image_meta_size:
-            self.style_embedder = nn.Embedding(1, embedding_dim)
-            extra_in_dim = 256 * 6 + embedding_dim + pooled_projection_dim
-        else:
-            extra_in_dim = pooled_projection_dim
-
-        self.extra_embedder = PixArtAlphaTextProjection(
-            in_features=extra_in_dim,
-            hidden_size=embedding_dim * 4,
-            out_features=embedding_dim,
-            act_fn="silu_fp32",
-        )
-
-    def forward(self, timestep, encoder_hidden_states, image_meta_size, style, hidden_dtype=None):
-        timesteps_proj = self.time_proj(timestep)
-        timesteps_emb = self.timestep_embedder(timesteps_proj.to(dtype=hidden_dtype))  # (N, 256)
-
-        # extra condition1: text
-        pooled_projections = self.pooler(encoder_hidden_states)  # (N, 1024)
-
-        if self.use_style_cond_and_image_meta_size:
-            # extra condition2: image meta size embdding
-            image_meta_size = self.size_proj(image_meta_size.view(-1))
-            image_meta_size = image_meta_size.to(dtype=hidden_dtype)
-            image_meta_size = image_meta_size.view(-1, 6 * 256)  # (N, 1536)
-
-            # extra condition3: style embedding
-            style_embedding = self.style_embedder(style)  # (N, embedding_dim)
-
-            # Concatenate all extra vectors
-            extra_cond = torch.cat([pooled_projections, image_meta_size, style_embedding], dim=1)
-        else:
-            extra_cond = torch.cat([pooled_projections], dim=1)
-
-        conditioning = timesteps_emb + self.extra_embedder(extra_cond)  # [B, D]
 
         return conditioning
